mds: split up mdstypes
mdstypes.h contains both MDS-exclusive and client-shared structs.
This PR splits it up into "mdstypes.h" and "sharedtypes.h".

Fixes: https://tracker.ceph.com/issues/3998
